backend/requirements.txt
# Core Framework
fastapi>=0.128.0
pydantic>=2.10.0
pydantic-settings==2.13.1
uvicorn[standard]>=0.34.0

# Database Async
asyncpg>=0.30.0

# Security
argon2-cffi>=23.1.0
python-jose[cryptography]
passlib[argon2]

# Graph DB
neo4j==5.28.1 

# Vector DB
pgvector==0.4.2

# HTTP Client(external API calls)
httpx==0.28.1

# utitlites
python-dotenv==1.2.1

# Local llm
ollama==0.6.1

# File uploads
python-multipart==0.0.20 

# Background tasks
celery[redis]==5.6.2
redis==5.2.1

